[Core] remove conversion to uint256 (149)

+++ src/core/contracts/libraries/AdaptiveFee.sol
@@ -24,7 +24,7 @@ library AdaptiveFee {
     uint256 sigm1 = sigmoid(volatility, config.gamma1, config.alpha1, config.beta1);
     uint256 sigm2 = sigmoid(volatility, config.gamma2, config.alpha2, config.beta2);
 
-    fee = uint256(config.baseFee) + sigmoid(volumePerLiquidity, config.volumeGamma, sigm1 + sigm2, config.volumeBeta);
+    fee = config.baseFee + sigmoid(volumePerLiquidity, config.volumeGamma, sigm1 + sigm2, config.volumeBeta);
   }
 
   function sigmoid(
